+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Added
 - Documented ``integral_deadzone`` in ``Control.pid()``.
+- Documented ``Motor.model``. This can be used to view the estimated motor
+  state and change its settings.
 
 ## 3.2.0 - 2022-12-20